delete unneeded components
